Add code comments

+++ SysMonService/Interfaces/ISettings.cs
@@ -1,8 +1,17 @@
 ﻿namespace SysMonService.Interfaces;
 
+/// <summary>
+/// Defines the configuration parameters for the system monitor service.
+/// </summary>
 public interface ISettings
 {
+    /// <summary>
+    /// Gets the identifier of the serial communication port (i.e. COM3)
+    /// </summary>
     public string? ComPort { get; }
     
-    public double? PollingRate { get; }
+    /// <summary>
+    /// Gets the interval in milliseconds at which data is transmitted.
+    /// </summary>
+    public int? PollingRate { get; }
 }
